requirements.txt
flask
requests
beautifulsoup4
openai
flask-cors
python-dotenv